Make nvJPEGDecoder cache global so that we can share it between instances
