+++ dali/pipeline/operators/decoder/CMakeLists.txt
@@ -12,6 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+add_subdirectory(cache)
+
 # Get all the source files and dump test files
 if (BUILD_NVJPEG)
   list(APPEND DALI_SRCS "${CMAKE_CURRENT_SOURCE_DIR}/nvjpeg_decoder.cu")
@@ -35,8 +37,6 @@ if (BUILD_TEST)
     "${CMAKE_CURRENT_SOURCE_DIR}/host_decoder_crop_test.cc"
     "${CMAKE_CURRENT_SOURCE_DIR}/host_decoder_slice_test.cc"
     "${CMAKE_CURRENT_SOURCE_DIR}/host_decoder_random_crop_test.cc"
-    "${CMAKE_CURRENT_SOURCE_DIR}/decoder_cache_blob_test.cc"
-    "${CMAKE_CURRENT_SOURCE_DIR}/decoder_cache_largest_only_test.cc"
     )
   set(DALI_TEST_SRCS ${DALI_TEST_SRCS} PARENT_SCOPE)
 endif()
