+++ dali/pipeline/operators/decoder/cache/CMakeLists.txt
@@ -1,0 +1,25 @@
+# Copyright (c) 2019, NVIDIA CORPORATION. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+file(GLOB tmp *.cc *.cu)
+file(GLOB tmp_test *_test.cc)
+remove(tmp "${tmp}" ${tmp_test})
+
+set(DALI_SRCS ${DALI_SRCS} ${tmp} PARENT_SCOPE)
+
+if (BUILD_TEST)
+  # get all the test srcs
+  file(GLOB tmp *_test.cc)
+  set(DALI_TEST_SRCS ${DALI_TEST_SRCS} ${tmp} PARENT_SCOPE)
+endif()
